Add files via upload
Using what we covered in week one, you are to create your first object by completing the following requirements.

1.    Create an object call Accounts. This is to be used in a banking system.
2.    Initialize the account with three data as inputs : Firstname, Lastname and initial deposit.
3.    Create 4 additional member functions: Deposit, Withdraw, Fee Calculations, interest
The fee calculation is going to be $10 per month if the total amount in the account is less than $1000. Interest is set to be at 3%.

You are to print the final values stored in your data objects and its member function.

All work is to be completed individually. You are welcome to make specific assumptions for your work.